pyproject.toml
[tool.poetry]
name = "tatm"
version = "0.1.0"
description = ""
authors = ["Ben Sabath <bsabath1@gmail.com>"]
readme = "README.md"
packages = [{include = "tatm", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
datasets = "^2.21.0"
tokenizers = "^0.20.0"
ray = {extras = ["default"], version = "^2.34.0"}
pyyaml = "^6.0.2"
omegaconf = "^2.3.0"
torch = "^2.4.1"
requests = "^2.32.3"

[tool.poetry.scripts]
tatm = "tatm:cli"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
sphinx = "^8.0.2"
myst-parser = "^4.0.0"
isort = "^5.13.2"


[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
isort = "^5.13.2"


[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^8.0.2"
myst-parser = "^4.0.0"
sphinx-rtd-theme = {git = "ssh://git@github.com/readthedocs/sphinx_rtd_theme.git", rev = "3.0.0rc1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"